Add the enum for the admin mode Modified the test with the new parameter Split the tests
