+++ LDAPLibrary/Enums/LDAPAdminMode.cs
@@ -1,0 +1,9 @@
+﻿namespace LDAPLibrary.Enums
+{
+    public enum LDAPAdminMode
+    {
+        Anonymous,
+        Admin,
+        NoAdmin
+    }
+}
